requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.10.0
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2
pydantic>=2.5.0
redis>=5.0.1
python-docx>=0.8.11
PyPDF2>=3.0.1
numpy>=1.24.3
python-multipart>=0.0.6
python-dotenv>=1.0.0